requirements.txt
tensorflow>=1.13.1
numpy>=1.16.2
Pillow>=5.4.1
opencv-python>=4.0.0.21
